Added a threaded CouchDB client.
It is meant to be used for sending commands to serial port, but this has
not yet been implemented. The client only checks for changes to a
document with id 'controller' and reports to stdout. For unsuccessful
connect attempts, there is a retry every 10 minutes. For disconnections
there is a cool-down period lasting a minute.